Moved extractTokenParts method into auth.passwordreset module
